Adicionando filtro

+++ src/SME.SGP.Infra/Dtos/Relatorios/NotasEConceitosFinais/FiltroRelatorioNotasEConceitosFinaisDto.cs
@@ -16,7 +16,7 @@ namespace SME.SGP.Infra.Dtos.Relatorios
         public string UsuarioRf { get; set; }
         public string UsuarioNome { get; set; }
         public CondicoesRelatorioNotasEConceitosFinais Condicao { get; set; }
-        public int ValorCondicao { get; set; }
+        public double ValorCondicao { get; set; }
         public TipoFormatoRelatorio TipoFormatoRelatorio { get; set; }
         public TipoNota TipoNota { get; set; }
     }
